Added Rock Paper Scissors  using Javascript
Created a simple Rock Paper Scissors  using Javascript. The HTML section consist of tow container, One for the computer, and the other for the player. Each section will contain the score and the icon. Under that container, we will put the rock, the paper, and the scissor buttons. Styled the HTML elements using CSS. In our JavaScript logic, we have created a variable for the computer score and the player score, then we selected our HTML elements. After that, we created an array that contains the rock, the paper, and the scissors icons in order to get them randomly from the array. Then we created the game function that contains all the logic, such as getting random icons from the array, putting them on the DOM, and the if statements for each situation(rock against paper, and etc). That function will run for each button clicked.